Minor renames and file shuffling
