+++ example/src/main/java/com/basistheory/android/example/util/PrettyPrintExtensions.kt
@@ -1,0 +1,8 @@
+package com.basistheory.android.example.util
+
+import com.google.gson.GsonBuilder
+
+fun Any.prettyPrintJson(): String {
+    val gson = GsonBuilder().setPrettyPrinting().create()
+    return gson.toJson(this)
+}
