+++ example/src/main/res/layout/fragment_custom_form.xml
@@ -4,13 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-
-    tools:context="com.basistheory.android.example.social_security_number.SocialSecurityNumberElement">
+    tools:context="com.basistheory.android.example.view.custom_form.CustomFormFragment">
 
     <ScrollView
+        android:background="@color/white"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/white">
+        android:layout_height="match_parent">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -18,20 +17,39 @@
             android:layout_margin="20dp"
             android:orientation="vertical">
 
+            <com.basistheory.android.view.TextElement
+                android:id="@+id/name"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:background="@drawable/rounded_edit_text"
+                android:padding="5dp"
+                bt:hint="Name"
+                bt:removeDefaultStyles="true"
+                bt:textColor="@color/gray_800" />
 
             <com.basistheory.android.view.TextElement
-                android:id="@+id/socialSecurityNumber"
+                android:id="@+id/phoneNumber"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="20dp"
                 android:background="@drawable/rounded_edit_text"
-                android:inputType="number"
                 android:padding="5dp"
-                bt:hint="Social Security Number"
-                bt:keyboardType="number"
-                bt:mask="###-##-####"
+                bt:hint="Phone Number"
                 bt:removeDefaultStyles="true"
                 bt:textColor="@color/gray_800" />
+
+            <com.basistheory.android.view.TextElement
+                android:id="@+id/orderNumber"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="20dp"
+                android:background="@drawable/rounded_edit_text"
+                android:padding="5dp"
+                bt:hint="Order Number"
+                bt:removeDefaultStyles="true"
+                bt:textColor="@color/gray_800"
+                bt:keyboardType="text"/>
+
 
             <LinearLayout
                 android:layout_width="wrap_content"
