+++ example/src/main/res/layout/fragment_social_security_number.xml
@@ -4,12 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.basistheory.android.example.text_element.TextElementFragment">
+    tools:context="com.basistheory.android.example.view.social_security_number.SocialSecurityNumberFragment">
 
     <ScrollView
-        android:background="@color/white"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/white">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -17,40 +17,19 @@
             android:layout_margin="20dp"
             android:orientation="vertical">
 
+
             <com.basistheory.android.view.TextElement
-                android:id="@+id/name"
+                android:id="@+id/socialSecurityNumber"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
                 android:background="@drawable/rounded_edit_text"
+                android:inputType="number"
                 android:padding="5dp"
-                bt:hint="Name"
+                bt:hint="Social Security Number"
+                bt:keyboardType="number"
+                bt:mask="###-##-####"
                 bt:removeDefaultStyles="true"
                 bt:textColor="@color/gray_800" />
-
-            <com.basistheory.android.view.TextElement
-                android:id="@+id/phoneNumber"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
-                android:background="@drawable/rounded_edit_text"
-                android:padding="5dp"
-                bt:hint="Phone Number"
-                bt:removeDefaultStyles="true"
-                bt:textColor="@color/gray_800" />
-
-            <com.basistheory.android.view.TextElement
-                android:id="@+id/orderNumber"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
-                android:background="@drawable/rounded_edit_text"
-                android:padding="5dp"
-                bt:hint="Order Number"
-                bt:removeDefaultStyles="true"
-                bt:textColor="@color/gray_800"
-                bt:keyboardType="text"/>
-
 
             <LinearLayout
                 android:layout_width="wrap_content"
