+++ example/src/main/res/values/strings.xml
@@ -6,8 +6,7 @@
     <string name="nav_header_subtitle">android.studio@android.com</string>
     <string name="nav_header_desc">Navigation header</string>
 
-    <string name="menu_home">Home</string>
-    <string name="menu_text_element">Custom Form Elements</string>
-    <string name="menu_social_security_number_element">Social Security Number</string>
-    <string name="menu_card_element">Card Elements</string>
+    <string name="menu_card">Collect Cards</string>
+    <string name="menu_social_security_number">Collect SSNs</string>
+    <string name="menu_custom_form">Custom Forms</string>
 </resources>
